requirements.txt
flask
pocketbase
flask_simple_crypt
python-dotenv
Flask-WTF
flask_ckeditor
